+++ R/add_annotation.R
@@ -36,6 +36,8 @@ get_cell_line_annotation <- function(
     )
   }
   
+  assert_cell_line_annotation(cell_line_annotation)
+  
   if (NROW(cell_line_annotation) > 0) {
     cmn <- cell_line_annotation[[clid]] %in% data[[clid]]
     cell_line_annotation <- cell_line_annotation[cmn, ]
@@ -91,7 +93,7 @@ annotate_dt_with_cell_line <- function(
     fill = "unknown"
 ) {
   checkmate::assert_data_table(data)
-  checkmate::assert_data_table(cell_line_annotation)
+  assert_cell_line_annotation(cell_line_annotation)
   
   cellline <- gDRutils::get_env_identifiers("cellline")
   add_clid <- gDRutils::get_header("add_clid")
@@ -104,6 +106,22 @@ annotate_dt_with_cell_line <- function(
     data[, (existing_cols) := NULL]
   }
   data <- cell_line_annotation[data, on = cellline]
+  
+  # Handle missing cell lines
+  missing_cell_lines <- setdiff(unique(data[[cellline]]), cell_line_annotation[[cellline]])
+  if (length(missing_cell_lines) > 0) {
+    missing_tbl_cell_lines <- data.table::data.table(
+      clid = missing_cell_lines,
+      CellLineName = missing_cell_lines,
+      Tissue = fill,
+      ReferenceDivisionTime = NA,
+      parental_identifier = fill,
+      subtype = fill
+    )
+    data.table::setnames(missing_tbl_cell_lines, names(cell_line_annotation))
+    data <- rbind(data[data[[cellline]] != missing_cell_lines],
+                  missing_tbl_cell_lines, fill = TRUE)
+  }
   (data)
 }
 
@@ -137,7 +155,6 @@ get_drug_annotation <- function(
   drug <- unlist(gDRutils::get_env_identifiers(c("drug", "drug2", "drug3"), simplify = FALSE))
   drug_ann_cols <- unlist(gDRutils::get_env_identifiers(c("drug", "drug_name", "drug_moa"), simplify = FALSE))
   
-  
   # Read the drug annotation data
   drug_annotation <- if (annotation_package == "gDRinternal") {
     eval(parse(text = paste0(annotation_package, "::get_drug_annotations")))()
@@ -146,6 +163,8 @@ get_drug_annotation <- function(
       system.file("annotation_data", fname, package = annotation_package), header = TRUE
     )
   }
+  
+  assert_drug_annotation(drug_annotation)
   
   untreated_tag <- gDRutils::get_env_identifiers("untreated_tag")
   all_data_drugs <- setdiff(gDRutils::remove_drug_batch(unlist(data[, intersect(names(data), drug), with = FALSE])),
@@ -193,7 +212,7 @@ annotate_dt_with_drug <- function(
     fill = "unknown"
 ) {
   checkmate::assert_data_table(data)
-  checkmate::assert_data_table(drug_annotation)
+  assert_drug_annotation(drug_annotation)
   
   drug <- unlist(gDRutils::get_env_identifiers(c("drug", "drug2", "drug3"), simplify = FALSE))
   untreated_tag <- gDRutils::get_env_identifiers("untreated_tag")
@@ -255,7 +274,6 @@ annotate_dt_with_drug <- function(
   (data)
 }
 
-
 #' Retrieve the drug annotation from the annotated dt input
 #'
 #' @param dt annotated data.table
@@ -300,7 +318,6 @@ get_drug_annotation_from_dt <- function(dt) {
   unique_dt[!unique_dt[[drug_cols[["drug"]]]] %in% gDRutils::get_env_identifiers("untreated_tag"), ]
 }
 
-
 #' Retrieve the cell line annotation from the annotated dt input
 #'
 #' @param dt annotated data.table
@@ -334,4 +351,169 @@ get_cellline_annotation_from_dt <- function(dt) {
   }
   
   unique(cell_dt)
+}
+
+#' annotate_se_with_drug
+#'
+#' Annotate SummarizedExperiment object with drug annotations
+#'
+#' @param se SummarizedExperiment object containing dose-response data
+#' @param drug_annotation data.table with drug annotations
+#' @param fill string indicating how unknown drugs should be filled in the DB
+#' @return SummarizedExperiment object with annotated drugs
+#' @keywords annotation
+#' @examples
+#' se <- SummarizedExperiment::SummarizedExperiment(
+#'   rowData = data.table::data.table(Gnumber = c("D1", "D2", "D3"))
+#' )
+#' drug_annotation <- get_drug_annotation(data.table::as.data.table(SummarizedExperiment::rowData(se)))
+#' annotated_se <- annotate_se_with_drug(se, drug_annotation)
+#' @export
+#'
+annotate_se_with_drug <- function(
+    se,
+    drug_annotation,
+    fill = "unknown"
+) {
+  checkmate::assert_class(se, "SummarizedExperiment")
+  assert_drug_annotation(drug_annotation)
+  checkmate::assert_string(fill)
+  
+  data <- data.table::as.data.table(SummarizedExperiment::rowData(se))
+  annotated_data <- annotate_dt_with_drug(data, drug_annotation, fill)
+  SummarizedExperiment::rowData(se) <- annotated_data
+  se
+}
+
+#' annotate_mae_with_drug
+#'
+#' Annotate MultiAssayExperiment object with drug annotations
+#'
+#' @param mae MultiAssayExperiment object containing dose-response data
+#' @param drug_annotation data.table with drug annotations
+#' @param fill string indicating how unknown drugs should be filled in the DB
+#' @return MultiAssayExperiment object with annotated drugs
+#' @keywords annotation
+#' @examples
+#' mae <- MultiAssayExperiment::MultiAssayExperiment(
+#'   experiments = list(exp1 = SummarizedExperiment::SummarizedExperiment(
+#'     rowData = data.table::data.table(Gnumber = c("D1", "D2", "D3"))
+#'   ))
+#' )
+#' drug_annotation <- get_drug_annotation(data.table::as.data.table(
+#'   SummarizedExperiment::rowData(
+#'     MultiAssayExperiment::experiments(mae)[[1]])))
+#' annotated_mae <- annotate_mae_with_drug(mae, drug_annotation)
+#' @export
+#'
+annotate_mae_with_drug <- function(
+    mae,
+    drug_annotation,
+    fill = "unknown"
+) {
+  checkmate::assert_class(mae, "MultiAssayExperiment")
+  assert_drug_annotation(drug_annotation)
+  checkmate::assert_string(fill)
+  
+  for (i in seq_along(MultiAssayExperiment::experiments(mae))) {
+    se <- MultiAssayExperiment::experiments(mae)[[i]]
+    MultiAssayExperiment::experiments(mae)[[i]] <- annotate_se_with_drug(se, drug_annotation, fill)
+  }
+  mae
+}
+
+#' annotate_se_with_cell_line
+#'
+#' Annotate SummarizedExperiment object with cell line annotations
+#'
+#' @param se SummarizedExperiment object containing dose-response data
+#' @param cell_line_annotation data.table with cell line annotations
+#' @param fill string indicating how unknown cell lines should be filled in the DB
+#' @return SummarizedExperiment object with annotated cell lines
+#' @keywords annotation
+#' @examples
+#' se <- SummarizedExperiment::SummarizedExperiment(
+#'   rowData = data.table::data.table(clid = c("CL1", "CL2", "CL3"))
+#' )
+#' cell_line_annotation <- get_cell_line_annotation(data.table::as.data.table(SummarizedExperiment::rowData(se)))
+#' annotated_se <- annotate_se_with_cell_line(se, cell_line_annotation)
+#' @export
+#'
+annotate_se_with_cell_line <- function(
+    se,
+    cell_line_annotation,
+    fill = "unknown"
+) {
+  checkmate::assert_class(se, "SummarizedExperiment")
+  assert_cell_line_annotation(cell_line_annotation)
+  checkmate::assert_string(fill)
+  
+  data <- data.table::as.data.table(SummarizedExperiment::rowData(se))
+  annotated_data <- annotate_dt_with_cell_line(data, cell_line_annotation, fill)
+  SummarizedExperiment::rowData(se) <- annotated_data
+  se
+}
+
+#' annotate_mae_with_cell_line
+#'
+#' Annotate MultiAssayExperiment object with cell line annotations
+#'
+#' @param mae MultiAssayExperiment object containing dose-response data
+#' @param cell_line_annotation data.table with cell line annotations
+#' @param fill string indicating how unknown cell lines should be filled in the DB
+#' @return MultiAssayExperiment object with annotated cell lines
+#' @keywords annotation
+#' @examples
+#' mae <- MultiAssayExperiment::MultiAssayExperiment(
+#'   experiments = list(exp1 = SummarizedExperiment::SummarizedExperiment(
+#'     rowData = data.table::data.table(clid = c("CL1", "CL2", "CL3"))
+#'   ))
+#' )
+#' cell_line_annotation <- get_cell_line_annotation(data.table::as.data.table(
+#'   SummarizedExperiment::rowData(
+#'     MultiAssayExperiment::experiments(mae)[[1]])))
+#' annotated_mae <- annotate_mae_with_cell_line(mae, cell_line_annotation)
+#' @export
+#'
+annotate_mae_with_cell_line <- function(
+    mae,
+    cell_line_annotation,
+    fill = "unknown"
+) {
+  checkmate::assert_class(mae, "MultiAssayExperiment")
+  assert_cell_line_annotation(cell_line_annotation)
+  checkmate::assert_string(fill)
+  
+  for (i in seq_along(MultiAssayExperiment::experiments(mae))) {
+    se <- MultiAssayExperiment::experiments(mae)[[i]]
+    MultiAssayExperiment::experiments(mae)[[i]] <- annotate_se_with_cell_line(se, cell_line_annotation, fill)
+  }
+  mae
+}
+
+#' Assert cell line annotation
+#'
+#' Validates that the cell line annotation data.table has the required columns.
+#'
+#' @param cell_line_annotation data.table with cell line annotations
+#' @keywords internal
+assert_cell_line_annotation <- function(cell_line_annotation) {
+  checkmate::assert_data_table(cell_line_annotation)
+  required_cell_line_cols <- unlist(gDRutils::get_env_identifiers(
+    c("cellline", "cellline_name", "cellline_tissue", "cellline_ref_div_time",
+      "cellline_parental_identifier", "cellline_subtype"), simplify = FALSE), use.names = FALSE)
+  checkmate::assert_names(names(cell_line_annotation), must.include = required_cell_line_cols)
+}
+
+#' Assert drug annotation
+#'
+#' Validates that the drug annotation data.table has the required columns.
+#'
+#' @param drug_annotation data.table with drug annotations
+#' @keywords internal
+assert_drug_annotation <- function(drug_annotation) {
+  checkmate::assert_data_table(drug_annotation)
+  required_drug_cols <- unlist(gDRutils::get_env_identifiers(
+    c("drug", "drug_name", "drug_moa"), simplify = FALSE), use.names = FALSE)
+  checkmate::assert_names(names(drug_annotation), must.include = required_drug_cols)
 }